menambahkan bulk delete

+++ app/Livewire/Employee.php
@@ -16,6 +16,7 @@ class Employee extends Component
     public $UpdateData = false;
     public $employee_id;
     public $katakunci;
+    public $employee_selected_id = [];
 
     public function store()
     {
@@ -73,19 +74,29 @@ class Employee extends Component
 
         $this->UpdateData = false;
         $this->employee_id = '';
+        $this->employee_selected_id = [];
     }
 
     public function delete()
     {
-        $id = $this->employee_id;
-        ModelsEmployee::find($id)->delete();
+        if($this->employee_id != ''){
+            $id = $this->employee_id;
+            ModelsEmployee::find($id)->delete();
+        }
+        if(count($this->employee_selected_id)){
+            for($x=0;$x<count($this->employee_selected_id);$x++){
+                ModelsEmployee::find($this->employee_selected_id[$x])->delete();
+            }
+        }
         session()->flash('message','Data berhasil di-delete');
         $this->clear();
     }
 
     public function delete_confirmation($id)
     {
-        $this->employee_id = $id;
+        if($id != ''){
+            $this->employee_id = $id;
+        }
     }
 
     public function render()
